POC for home page with live updates
HUGE lift on this from `Phoenix.PubSub`.

I'm not sold on the `with` statement that I used in the `Groups` module,
but I think that a `with` is probably the way to go, and this could just
use some refactoring.

There's also an unnecessary `handle_info` callback in the `HomeLive`
module, which I was using to help me debug. I've left it in, since it
may be useful as I continue to play with this.

I just don't have the time/ attention for that right now, at 2130 on a
saturday :) and I would rather get the code up to main than leave it
around to wait for a refactor.