Add imports for fixtures in `iex`
Without setting the `MIX_ENV` to "test", the fixture modules are not
compiled, and executing the start-up `.iex.exs` fails.

I thought that I could fix this by using a short-circuit conditional,
but I was wrong. Even though it's not doing anything right now, I am
going to leave it in since it has the same functionality as before and
will remind me of my effort/ previous avenue of investigation.