Add functions to add and remove `GroupTeam` associations
Deleting from a table with no primary key appears to be a bit tricky.

Calling `Repo.delete/2` results in a raised exception, because `delete/2` expects the table to have a PK.

Googling for this led me to a google group page with a back-and-forth
between José Valim and Justin Schneck! and therein lay the answer.

https://groups.google.com/g/elixir-ecto/c/sEbaz2j9wX0/m/NrOs12LSJ10J