Rework the `add_x/2` funcs for `Groups` and `Teams`
Calling the `add_x/2` function multiple times would create multiple rows
in the join table. Not the end of the world, but if I want to make use
of the `timestamps` on my join tables then that would get pretty tricky,
unless each join record had extra meta-data, etc...

There is a growing amount of `Ecto.Query` logic in the `Groups` and
`Teams` context modules now that is related to `Joins`.
This should probably be moved out into a `JoinsQuery` module or similar
to keep the noise out of the contexts.

The `Joins` modules _might_ be a better place for the `add_x/2`
functions also, but I will have to see how they feel in their current
spots to decide.